update: 角度と目盛りの対応表を追加しました

+++ include/Entity/Entity.hpp
@@ -22,7 +22,7 @@ namespace Game {
         Core::AudioHandle timeSound_;
         float animeScale = 1;
 
-        float timeValue = 0;
+        int timeValue = 0;
         Core::Point localP;
         Core::Point virtualP;
         Core::Matrix2x2 rotateMatrix;
